src/main/java/com/lsh/service/impl/CategoryServiceImpl.java
package com.lsh.service.impl;

import com.lsh.dao.ProductCategoryDao;
import com.lsh.dataobject.ProductCategory;
import com.lsh.service.CatergoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * @ClassName CategoryServiceImpl
 * @Description: 类目
 * @Author lsh
 * @Date 2018/10/7 13:08
 * @Version
 */
@Service
public class CategoryServiceImpl implements CatergoryService {
    @Autowired
    private ProductCategoryDao productCategoryDao;
    @Override
    public ProductCategory findOne(Integer categoryId) {
        return productCategoryDao.findOne(categoryId);
    }

    @Override
    public List<ProductCategory> findAll() {
        return productCategoryDao.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return productCategoryDao.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return productCategoryDao.save(productCategory);
    }
}
